+++ src/contents/sounds.tsx
@@ -11,11 +11,17 @@ export default function SoundsContent(){
      return (
           <WindowWrapper title={t("title")} className="mt-8 mb-12 flex justify-center items-center flex-col gap-2 w-full">
                <div className="w-full grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 gap-1.5">
-                    {sounds.map(sound=>(
-                         <Suspense key={sound.id} fallback={<SoundCardLoader/>}>
+                    <Suspense fallback={(
+                         <>
+                         {Array.from({length: 10}).map((_,i)=>(
+                              <SoundCardLoader key={i+1}/>
+                         ))}
+                         </>
+                    )}>
+                         {sounds.map(sound=>(
                               <SoundCard data={sound}/>
-                         </Suspense>
-                    ))}
+                         ))}
+                    </Suspense>
                </div>
           </WindowWrapper>
      )